backend/requirements.txt
# Flask and core dependencies
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.14.0

# Firebase Admin SDK for token verification
firebase-admin==7.0.0

# Security
PyJWT==2.10.1
bcrypt==4.2.1
Werkzeug==3.1.3

# Validation
pydantic==2.10.3

# Development tools
gunicorn==23.0.0
